Refactoring
Moving concrete implementations to infrastructure project
