+++ OperationsSolver.Test/Infrastructure/IO/JsonReaderTest.cs
