+++ OperationsSolver.Test/Infrastructure/Operations/SumTest.cs
@@ -1,4 +1,4 @@
-﻿using OperationsSolver.Application.Operations;
+﻿using OperationsSolver.Infrastructure.Operations;
 
 namespace OperationsSolver.Test.Logic.Operations
 {
